Move jmac package in ee-security to authentication
