+++ appserver/security/core-ee/src/main/java/com/sun/enterprise/security/ee/authentication/jakarta/WebServicesDelegate.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package com.sun.enterprise.security.ee.jmac;
+package com.sun.enterprise.security.ee.authentication.jakarta;
 
 import com.sun.enterprise.deployment.ServiceReferenceDescriptor;
 import com.sun.enterprise.deployment.runtime.common.MessageSecurityBindingDescriptor;
