+++ appserver/security/core-ee/src/main/java/com/sun/enterprise/security/ee/authentication/jakarta/callback/ClientContainerCallbackHandler.java
@@ -20,7 +20,7 @@
  * Created on September 14, 2004, 12:20 PM
  */
 
-package com.sun.enterprise.security.ee.jmac.callback;
+package com.sun.enterprise.security.ee.authentication.jakarta.callback;
 
 import java.io.IOException;
