--- appserver/security/core-ee/src/main/java/com/sun/enterprise/security/ee/jmac/config/GFAuthConfigFactory.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0, which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the
- * Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
- * version 2 with the GNU Classpath Exception, which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- */
-package com.sun.enterprise.security.ee.jmac.config;
-
-import org.glassfish.epicyro.config.factory.file.AuthConfigFileFactory;
-
-/**
- * Temporarily kept around, since there are external references to this class.
- *
- * <p>
- * Should be removed in the next major version of GlassFish.
- */
-@Deprecated(forRemoval = true)
-public class GFAuthConfigFactory extends AuthConfigFileFactory {
-
-}
