+++ appserver/security/webservices.security/src/main/java/com/sun/enterprise/security/webservices/client/ClientSecurityPipe.java
@@ -30,7 +30,7 @@ import static jakarta.security.auth.message.AuthStatus.SEND_SUCCESS;
 import static java.util.logging.Level.FINE;
 import static java.util.logging.Level.SEVERE;
 
-import com.sun.enterprise.security.ee.jmac.callback.ClientContainerCallbackHandler;
+import com.sun.enterprise.security.ee.authentication.jakarta.callback.ClientContainerCallbackHandler;
 import com.sun.enterprise.security.webservices.LogUtils;
 import com.sun.enterprise.security.webservices.SoapAuthenticationService;
 import com.sun.enterprise.util.LocalStringManagerImpl;
